refactor BudgetMonthCountContext to tsx (#1722)
* refactor BudgetMonthCountContext to tsx

* add release notes
