+++ upcoming-release-notes/1722.md
@@ -1,0 +1,6 @@
+---
+category: Maintenance
+authors: [Jod929]
+---
+
+Refactor budget/BudgetMonthCountContext to tsx.
